changed test to be gtest
